add MustBeIndirect flag to PdfObject; removed document from PdfStream and most object wrappers constructors
From now on pdf streams are direct on creation and are transformed into indirects either by explicit call of makeIndirect() method or when the stream is flushed. Also, most of the object wrappers are now direct on creation via any constructor, and they are supposed to be transformed into indirect objects only explicitly.

DEVSIX-207