Improve lineheight calculation
If we find WinAscender and WinDescender in the font and they are not equal to TypoAscender and TypoDescender correspondingly, we use their values to calculate line height. Otherwise, we use TypoAscender and TypoDescender values scaled by a factor of 1.2 (to fit diacritic sybmols in line area). Multiplied leading is calculated based on the ascender/descender values above scaled by font size and the specified multiplier. It does not therefore equal to the fontSize * leadingScaleFactor in the resultant pdf and may be more or less than that value. If the multiplied leading factor is equal to one, the lines will be placed side-by-side.

DEVSIX-264