Remove parser.GraphicState class, CanvasGraphicState class and parser package refactoring
Moved Path and connected to it classes from parser package to geom package.
CanvasGraphicState fields now have default values according to the spec. Also CanvasGraphicState now tracks the ctm.
Instead of the parser.GraphicState there is a new internal class ParserGraphicState which extedns the CanvasGraphicState and encapsulates the logic of clipping path tracking.