Fix wrong order of canvas operators applied for border painting, mark table borders as artifact
As stated in PDF Reference, 4.1 Grpahics Objects: "... a PDF content stream is not a program to be interpreted; rather, it is a static description of a sequence of graphics objects.". So it means that strictly saying those objects should not overlap each other. Path object is one of these graphics objects and it is described in spec (FIGURE 4.1 Graphics objects) that it only contains path construction operators. Even though most of the pdf applications accept such content streams with other operators in path objects, it is wrong to write such content streams.

IEA-8